requirements.txt
numpy>=1.20.0
open3d>=0.15.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
torch>=1.10.0
torchvision>=0.11.0
pandas>=1.3.0
tqdm>=4.62.0
pyyaml>=6.0
pytest>=6.2.5
jupyter>=1.0.0 